+++ enterprise-java-project/src/main/java/com/project/service/request/RunnerRequest.java
@@ -3,14 +3,14 @@ package com.project.service.request;
 import com.project.enums.RunnerStatusEnum;
 
 public class RunnerRequest extends UserRequest {
-    private double deliveryFees;
+    private float deliveryFees;
     private RunnerStatusEnum runnerStatus;
 
-    public double getDeliveryFees() {
+    public float getDeliveryFees() {
         return deliveryFees;
     }
 
-    public void setDeliveryFees(double deliveryFees) {
+    public void setDeliveryFees(float deliveryFees) {
         this.deliveryFees = deliveryFees;
     }
